fix touch in apps, rename messages test file, increase coverage, refactor a bit more
