+++ src/modules/utils.ts
@@ -1,0 +1,3 @@
+export const clamp = (number: number, min: number, max: number) => {
+  return Math.max(min, Math.min(number, max))
+}
